rename: move file from inadequate directory
